requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
cloudinary==1.23.0
dj-database-url==0.5.0
Django==3.1.3
django-appconf==1.0.4
django-cloudinary-storage==0.3.0
django-comments-xtd==2.8.1
django-contrib-comments==1.9.2
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-image-cropping==1.5.0
django-model-utils==4.1.1
django-quill-editor==0.1.16
django-stdimage==5.1.1
django-tinymce4-lite==1.8.0
djangorestframework==3.12.2
docutils==0.16
easy-thumbnails==2.7
gunicorn==20.0.4
idna==2.10
jsmin==2.2.2
jsonfield==3.1.0
Pillow==8.0.1
progressbar2==3.53.1
psycopg2==2.8.6
python-memcached==1.59
python-utils==2.4.0
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
urllib3==1.26.2
whitenoise==5.2.0
